// data/text-data/main.enum.ts
export enum MainTextData {
    INVALID_EMAIL = "Invalid email or password. Please try again or click ‘Forgot your password?’ link to reset it",
    ID_PLACEHOLDER = "ID no.",
    PASSWORD_PLACEHOLDER = "Password",
    EMAIL_PLACEHOLDER = "Email",
    EMAIL_CLICK_PLACEHOLDER = "email@address.com",
    INVALID_ID = "Invalid ID or password. Please try again or click 'Forgot your password?' link to reset it",
}